Refactored packaging architecture for entire project
