+++ WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/MainController.java
@@ -1,0 +1,18 @@
+package com.webshopapp.adminpanel;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainController {
+
+    @GetMapping
+    public String viewHomePage() {
+        return "index";
+    }
+
+    @GetMapping("/login")
+    public String LoginPage() {
+        return "login";
+    }
+}
