+++ WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/category/dto/CategoryDTO.java
@@ -1,0 +1,14 @@
+package com.webshopapp.adminpanel.category.dto;
+
+
+import lombok.*;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class CategoryDTO {
+
+    private Integer id;
+    private String name;
+
+}
