--- WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/controller/MainController.java
@@ -1,18 +1,0 @@
-package com.webshopapp.adminpanel.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-
-    @GetMapping
-    public String viewHomePage() {
-        return "index";
-    }
-
-    @GetMapping("/login")
-    public String LoginPage() {
-        return "login";
-    }
-}
