--- WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/controller/file/ApplicationConfig.java
@@ -1,27 +1,0 @@
-package com.webshopapp.adminpanel.controller.file;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.nio.file.Path;
-import java.nio.file.Paths;
-
-@Configuration
-public class ApplicationConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        imagesDirectory("category-images", registry);
-        imagesDirectory("brand-images", registry);
-        imagesDirectory("product-images", registry);
-    }
-
-    private void imagesDirectory(String pathPattern, ResourceHandlerRegistry registry) {
-        Path path = Paths.get(pathPattern);
-        String absolutePath = path.toFile().getAbsolutePath();
-        String logicalPath = "/" + pathPattern + "/**";
-        registry.addResourceHandler(logicalPath)
-                .addResourceLocations("file:/" + absolutePath + "/");
-    }
-}
