--- WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/dto/CategoryDTO.java
@@ -1,14 +1,0 @@
-package com.webshopapp.adminpanel.dto;
-
-
-import lombok.*;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class CategoryDTO {
-
-    private Integer id;
-    private String name;
-
-}
