--- WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/repository/RoleRepository.java
@@ -1,9 +1,0 @@
-package com.webshopapp.adminpanel.repository;
-
-
-import com.webshopapp.common.entity.role.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-
-}
