+++ WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/role/RoleRepository.java
@@ -1,0 +1,9 @@
+package com.webshopapp.adminpanel.role;
+
+
+import com.webshopapp.common.entity.role.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoleRepository extends JpaRepository<Role, Integer> {
+
+}
