+++ WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/security/SecurityConfig.java
@@ -1,7 +1,6 @@
 package com.webshopapp.adminpanel.security;
 
 
-import com.webshopapp.adminpanel.security.user.SecurityUserDetailsService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
