+++ WebParent/AdminPanel/src/main/java/com/webshopapp/adminpanel/security/SecurityUserDetails.java
@@ -1,0 +1,66 @@
+package com.webshopapp.adminpanel.security;
+
+
+import com.webshopapp.common.entity.role.Role;
+import com.webshopapp.common.entity.user.User;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Set;
+
+public class SecurityUserDetails implements UserDetails {
+
+    private User user;
+
+    public SecurityUserDetails(User user) {
+        this.user = user;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        Set<Role> roles = this.user.getRoles();
+        List<SimpleGrantedAuthority> authorities = new ArrayList<>();
+        for (Role role : roles) {
+            authorities.add(new SimpleGrantedAuthority(role.getName()));
+        }
+        return authorities;
+    }
+
+    @Override
+    public String getPassword() {
+        return this.user.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return this.user.getEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return this.user.isStatus();
+    }
+
+    public String getAccountName() {
+        return this.user.getFirstName() + " " + this.user.getLastName();
+    }
+}
