+++ WebParent/AdminPanel/src/test/java/com/webshopapp/adminpanel/UserRepositoryTests.java
@@ -3,7 +3,7 @@ package com.webshopapp.adminpanel;
 
 import com.webshopapp.common.entity.role.Role;
 import com.webshopapp.common.entity.user.User;
-import com.webshopapp.adminpanel.repository.UserRepository;
+import com.webshopapp.adminpanel.user.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
