+++ WebParent/CustomerPanel/src/main/resources/templates/product/product_detail.html
@@ -30,7 +30,7 @@
             </div>
 
             <div class="mt-2">
-                <span class="badge badge-info" style="font-size: 20px">
+                <span class="badge badge-success" style="font-size: 20px">
                     <b>Price:</b> $ [[${product.price}]]
                 </span>
             </div>
